refactor: 유효기간 커스텀 훅을 통한 컴포넌트 책임분리 > 상태전이 도메인로직 분리
